SFEperl-file-copy-recursive: package for perl-522 is added and for perl-520 is obsolete
--HG--
branch : spec_SFEperl-file-copy-recursive

+++ specs/SFEperl-file-copy-recursive.spec
@@ -4,7 +4,7 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define include_executable 0
 
 %define cpan_name File-Copy-Recursive
@@ -29,6 +29,7 @@ Perl extension for recursively copying files and directories
 IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Perl extension for recursively copying files and directories
 BuildRequires:    runtime/perl-584 = *
+BuildRequires:    library/perl-5/pathtools-584
 Requires:         runtime/perl-584 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-584
@@ -42,6 +43,7 @@ Perl extension for recursively copying files and directories
 IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Perl extension for recursively copying files and directories
 BuildRequires:    runtime/perl-510 = *
+BuildRequires:    library/perl-5/pathtools-510
 Requires:         runtime/perl-510 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-510
@@ -55,6 +57,7 @@ Perl extension for recursively copying files and directories
 IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Perl extension for recursively copying files and directories
 BuildRequires:    runtime/perl-512 = *
+BuildRequires:    library/perl-5/pathtools-512
 Requires:         runtime/perl-512 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-512
@@ -68,6 +71,8 @@ Perl extension for recursively copying files and directories
 IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Perl extension for recursively copying files and directories
 BuildRequires:    runtime/perl-516 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+BuildRequires:    library/perl-5/pathtools-516
 Requires:         runtime/perl-516 = *
 Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/pathtools-516
@@ -76,23 +81,24 @@ Requires:         library/perl-5/pathtools-516
 Perl extension for recursively copying files and directories
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Perl extension for recursively copying files and directories
-BuildRequires:    runtime/perl-520 = *
-Requires:         runtime/perl-520 = *
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/pathtools-522
+Requires:         runtime/perl-522 = *
 Requires:         library/perl-5/%{ips_cpan_name}
-Requires:         library/perl-5/pathtools-520
+Requires:         library/perl-5/pathtools-522
 
-%description 520
+%description 522
 Perl extension for recursively copying files and directories
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -105,8 +111,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ x${test} = 'xwithout_test' ] || make test
+
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null
+    if [ $? -eq 0 ]
+    then
+        make CC='cc -m32' LD='cc -m32'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m32' LD='cc -m32'
+    else
+        make CC='cc -m64' LD='cc -m64'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m64' LD='cc -m64'
+    fi
+
     make pure_install
 }
 
@@ -121,7 +136,7 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ x${test} = 'xwithout_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build clean
 }
@@ -156,7 +171,11 @@ modify_man_dir() {
             mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
             rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
         fi
-        rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
     fi
 }
 
@@ -191,8 +210,8 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
@@ -253,18 +272,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Wed Apr 05 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added and for perl-520 is obsolete
 * Sat Nov 14 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - build packages for perl-510, perl-516 and perl-520
 * Sun Jun 10 2012 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
